Launch last three activities
Created the TaskDetailActivity and made it launch when you press a task.
Also made the ClassmatesActivity and WorkManagerActivity that Phil made
earlier launch when you press a menu item. Tried to put RecyclerViews
with dummy data in those last two activities but the app can't get the
RecyclerViews with findViewById