Sharing tasks with classmtes via email
By long clicking on a task in work activity, you can email your
classmates the key for that task. The receiver(s) can copy the key from
the email and paste it in a dialog launched from a toolbar item. The app
will then use that key to find the sender's task, copy the data, and
create an exact copy in your class!